Update Settings.lua
(Very small fix to a "typo" in my code. Wouldn't matter much, but wouldn't look nice)

Forgot to mention:
devConsoleButton.MouseButton1Click has been edited that whenever you press the Log-button in
the Help section of the menu, it'll resume the game, aka hide the menu.
Currently whenever you press the button, it exists the help section to the main menu.