Update RbxGui.lua
Added a check for insertFrame.Button in the setInsertButtonImageBehavior function. In some weird cases, the button is occasionally absent and causes the output to be flooded with "Button is not a valid member of Frame"
This can be replicated in the Stamper Tool plugin if you rapidly try to change sets.