Change OOP from closure-based to metatable based
This update does a few things. It may seem like a lot of code, but it’s
really just refactoring variables.

* Removes the :Initialize() block, which is a bad OOP practice
* Starts to avoid mixing logic and instantiation.
     * Note several more changes will be made soon to prevent
construction inside of the window itself
     * This is considered good OOP design
 * Allows the extension of the ChatWindow through inheritance

Several more changes will be needed to entirely decouple the chat
window from globals, but this is a good start.

I understand if ROBLOX doesn’t actually want this change, but as far as
code quality goes, this is a good step forwards.