Properly use the new API
creatorFlagEnabled is always true due to the API being enabled.
This results in isCreator being true too, acting as if the player owns the current place.
Apparently, actually using the new API wasn't implemented here.

This commit should switch from checking the game's CreatorId to the new API, if it's (still) disabled.
Proper error handling is included. When the new method fails, it'll resort to the old method again.