Lua scrollbar + minute alignment adjustments + various fixes
- Fixed clipping + touch controls + drag controls
- Removed the line at the top (per tara's request)
- Fixed tweening menus to the left.  Previously when you moved to the
left menu the old one would tween out in the left direction (the same
direction the new one is tweening in from).  Now it tweens out in the
right direction.
- Fixed slider tick image states.  Previously if you turned a slider
down, dragged out of the frame to deselect it, then dragged back up, the
newly-appearing ticks would stay in the selected state while the rest of
the slider stayed deselected.
- Fixed arrows on sliders not updating if SetMinStep was used to make
the min step the current step.  This would result in setting mouse
sensitivity to 1, rejoining the game, reopening the menu, and the down
arrow would be there even though you were at the lowest setting.
- Fixed mouse over button states activating when the fullscreen drop
down menu is open