Added Gamepad Camera Zooming (via Thumbstick)
I decided to try and implement the mouse-wheel zooming feature into the RootCamera module.

When you hold down the right thumbstick, the camera will zoom forward if you are pushing it forward, and backwards if you push it backwards. The camera's pan position is multiplied by 0.5x0.1 while the zooming occurs.
The zoom acceleration depends on the current zoom of the camera itself (currently it clamps to 10 <= currentZoom <= 400)
If you press on the right analog stick, and then immediately release, the camera will switch between first and third person as it currently does.

Give it a try, let me know what you think, and if there are any severe problems I should address, let me know :)
I hope you guys consider implementing it!