Fixed hurt overlay showing when health CoreGui was disabled
Whenever the player takes a significant amount of damage, the topbar
CoreScript flashes a hurt overlay. The only problem was that even if the
health CoreGui was disabled, this hurt overlay still flashed. This adds
an extra check to make sure the hurt overlay is not used when the health
CoreGui is disabled.