Update HealthScript.lua
Changed the way the script waits for game.Players.LocalPlayer.
Removed a wait until game:GetService("Players"), as it is contradicting to have a wait for players and a wait for game.Players.LocalPlayer. When you could just have a wait until game.Players.LocalPlayer and accomplish the same thing.