CLI-10852: Add Advanced Mouse Sensitivity options to Client
This is the UI portion of the changes.  The UI will have additional fields for
a basic/advanced mode selector.  In advanced mode, the UI will have an option
to set the mouse sensitivity for 3d use-cases to a floating point value.  The
display will round to three decimal places.  The option to invert the mouse in
3d use-cases is also provided.  When using the advanced mode, the setting for
basic mode is ignored.

Internally, these changes use my hack week project which included independent
x, y scaling for the "LockCenter" and the "LockCurrentPosition" MouseBehavior.
The values are saved to GlobalBasicSettings_13.xml and loaded during the
application launch.

The intent is to be more friendly to users who prefer and inverted mouse,who
use low mouse sensitivity, or who use high-dpi gaming mice.