CLI-11421 Fixed Mobile controls disappearing when character is loaded in.
Because CharacterRemoving is never called when
Players.CharacterAutoLoads = false on the initial load in, Control State
would never be set to nil, causing the UI to not enable when the
character loaded in. This fix moves the code from the CharacterRemoving
event connection to the CharacterAdded one, which has no effect on the
code other than accommodating CharacterAutoLoads.