CLI-11421 Fixed Mobile controls disappearing when character is loaded in. (#416)
* CLI-10871 fixed radial menu toggle

made it so that the game has to be loaded in to activate the radial
menu, using game.Loaded event.

* CLI-10871 fixed radial menu toggle

Slightly cleaner fix for the code than previously. Uses game.Loaded
event to detect when the game starts to turn the radial menu togglable.

* CLI-11421 Fixed Mobile controls disappearing when character is loaded in.

Because CharacterRemoving is never called when
Players.CharacterAutoLoads = false on the initial load in, Control State
would never be set to nil, causing the UI to not enable when the
character loaded in. This fix moves the code from the CharacterRemoving
event connection to the CharacterAdded one, which has no effect on the
code other than accommodating CharacterAutoLoads.

* CLI-2725 Fixed blocked users being able to communicate with players

Previously, GetBlockedPlayersAsync was returning a table with the key
being incremental, with the value being a blocked UserId, while the
isBlocked function expected a dictionary of the UserIds. This fix
changes what  GetBlockedPlayersAsync returns to a dictionary that fits
what the function is expecting.

* CLI-2725 Fixed blocked users being able to communicate with players

Updated to remove debug code

* CLI-11421 Fixed Mobile controls disappearing when character is loaded in.

Readded CharacterRemoving connection