Updated GetFontHeight to work with all Fonts/Sizes
Right now, certain Fonts and FontSizes don't work with this function, because they haven't been registered.
I rewrote the function, so that it actually computes the height of the font using a couple of basic rules, rather than just returning from a preset list of returns. This will ensure support with future fonts.
I have tested this, and confirmed that it returns the exact same results that it used to return.