CLI-10725 Buttons should be smaller in the UI for mobile (#424)
* CLI-10725 Buttons should be smaller in the UI for mobile

Made the mobile Jump and context action buttons the same size as the interior thumbstick button (35x35 pixels on small devices, 45,45 on larger ones) and then positioned them accordingly to have roughly the same position on the other side of the screen.

* CLI-10725 Buttons should be smaller in the UI for mobile

Made the mobile Jump and context action buttons the same size as the interior thumbstick button (35x35 pixels on small devices, 45,45 on larger ones) and then positioned them accordingly to have roughly the same position on the other side of the screen.

* CLI-11443 ContextActionService:BindAction() not creating button on Mobile

When the PlayerGui refreshes, buttonFrame gets parented to nil, causing
all buttons displayed within it to not show. This fix reparents it to
buttonScreenGui when its parent doesn't exist, which basically happens
when createButton is called for the first time in a character's life.

* Readjusted Jump button position

The jump button should now be level with the thumbstick on all devices.