CLIMOB-2466 Fixed other Movement Modes not working after spawning wit… (#431)
* CLIMOB-2466 Fixed other Movement Modes not working after spawning with ClickToMove enabled

Because of how ClickToMove isn't a module of the control script, the necessary code for connecting the movement style changes wasn't running, due to it not getting a module. This fix makes it so that these events are called outside of that if statement. So far this change works fine on my mobile devices; but I'll look into the ClickToMove module and Camera Script to make sure there aren't any unforeseen issues.

* Moved jump module lines out of if statement

in the RefreshControlStyle function.