Add Keyboard ControlModule refresh when (Dev)ComputerMovementMode changes (#444)
This is a fix for the following bug:
When changing the ComputerMovementMode or DevComputerMovementMode the Keyboard ControlModule does not update. Currently this bug is only observed when switching to/from the Scriptable DevComputerMovementMode. This does not happen with the Touch ControlModule because it listens for changes to (Dev)ComputerMovementMode.

This fix mirrors over the event listening and module refreshing code from the Touch ControlModule to the Keyboard ControlModule in order to remain consistent.
This issue was originally pointed out at http://devforum.roblox.com/t/changing-devcomputermovementmode-or-devtouchmovementmode-sometimes-doesnt-activate-instantly/26844