Changes to the Chat UI
Includes a few features that Adam requested, such as chat filling in from the top down, removal of the "Please chat '/?'..." message when it first launches, and I changed the Chat container to a Frame from a TextButton on all platforms but VR, because it had no purpose being a TextButton. Lastly I made it not Active, so that player made UI behind the chat can be interacted with.