New Chat only create labels for visible channels. (#556)
* Refactor MessageLabelCreator to reduce duplicated code.

* Fix channel bar scrolling.

The right arrow did not appear due to ScrollChannelsFrame being
unnecssarly locked when the channels are added. If the direction is 0,
tweening is not necessary.

* Refactor chat to only create text labels for visible channels.

This greatly improves performance when there are a lot of channels.

* New Chat object pool.

Use an object pool to avoid constantly creating new frames and text
labels.

* Refactor new chat so that code is not contained in giant strings.

This makes the code easier to read.

* Fix multiple channels message filtering bug

When you switched from one channel to another, the fact that a message
was filtered would be forgotten.

* Remove debugging code

* Move object pool to common.

* Move settings to chat settings file.

Some settings were not in the setttings table. Moved them there so they
are easy to find.

* Create common module to force new chat on.

Otherwise both the ServerStarterScript and the ChatSelector have to be
edited.