CLI-3017 Mobile Jump Button Should Disappear If Jumping Disabled (#558)
* touch jump changes

made MasterControl:GetHumanoid public, and used this in TouchJump to get
the humanoid. When the JumpPower property changes,  if it's less than 1,
it hides the button, if its greater, it'll display it.

* Added check for default disabled jump

In the case the jump power is set before jump button initializes i added
this check.

* should work on respawn, and some other changes

I reworked some of the code to make it more organized, buttons state
changing should work on respawn, and made it so if the jump button gets
disabled externally, it knows better than to reenable it without
checking that first.