New Chat Message Creator modules. (#567)
* Refactor MessageLabelCreator to reduce duplicated code.

* Fix channel bar scrolling.

The right arrow did not appear due to ScrollChannelsFrame being
unnecssarly locked when the channels are added. If the direction is 0,
tweening is not necessary.

* Refactor chat to only create text labels for visible channels.

This greatly improves performance when there are a lot of channels.

* New Chat object pool.

Use an object pool to avoid constantly creating new frames and text
labels.

* Refactor new chat so that code is not contained in giant strings.

This makes the code easier to read.

* Fix multiple channels message filtering bug

When you switched from one channel to another, the fact that a message
was filtered would be forgotten.

* Remove debugging code

* Move object pool to common.

* Move settings to chat settings file.

Some settings were not in the setttings table. Moved them there so they
are easy to find.

* Create common module to force new chat on.

Otherwise both the ServerStarterScript and the ChatSelector have to be
edited.

* Fix chat text fading out

* Output error if SetCore CoreGuiChatConnections fails

* Fix disappearing text bug.

* Move ChatSettings to seperate ClientChatModules.

This allows easy editing without having to edit the whole chat and lose
the ability for it to update.

* Fix filtering for /me messages.

* New Chat fix blue text bug.

Change message mode to a text label as sometimes it was possible to
mistakenly click into the text box. There did not seem to be a need for
this to be a text box.

* New chat message creation modules.

Move message label creation to client chat modules.

* Assorted new chat changes

* Add basic instructions for creating a Message Creator Module

* Return object from message creators, add headers.