Remove heightScalingFlag  (#580)
* CLI-13224 remove unused useHttps parameter in RbxHttpApiService requests

* Remove old flag, fix for InUse never being set back to false

Also ensure that Dialog GUI is only shown if the Dialog is a descendant
of Workspace

* Fix dialogs in case where player leaves with an active conversation

* Warning on empty InitialPrompt

* Flag changes to ServerStarterScript

* Add scaling for camera height

* Remove heightScalingFlag

Adding new flags will cause issues later when the flag is removed and
developers have copied the PlayerScripts. Removing the flag later will
make the flag false for any copies of the script which are not updated.