Fix following icon (#679)
Fix all follwing icons being overriden when a new player joins the game. This happens because the following status of a new player is sent in the same format as the initial data. Thus, the script incorrectly attempts to set the icon to nil because no data is sent for already existing players.