Prevent follow notifications spam (#711)
There is a fix in engine for friend request notification spam but follow
notifications are not covered by this fix.

I think a reasonable approach is to limit the amount of follow
notifications to a set amount between any two players and not send any
more after that. We could also make the max number a fast int if we want
the option to configure it later.