Fix centering of alert dialog on mobile
The dialog shown after submitting a report was off-centered to the
right. The code assumed that the AlertViewBacking had an absolute size
and not a relative sizee. This caused the backing to start in the middle
of the screen and extend offscreen since it has a horizontal scale of
0.8 on mobile. The code has been corrected to handle both scale and
absolute sizes.