Check that sentNotificationMap exists before indexing it (#735)
* CLI-14718: Fix reset button customization on small touch screen devices

Commit also includes some other small style tweaks.

* Add throttle for receiving friend notifications.

I'm hoping this will mitigate some of the friend request spam issues but
it's not a full solution. We need to fix this on the server.

* Add FInt for friend request throttle

* Prevent follow notifications spam

There is a fix in engine for friend request notification spam but follow
notifications are not covered by this fix.

I think a reasonable approach is to limit the amount of follow
notifications to a set amount between any two players and not send any
more after that. We could also make the max number a fast int if we want
the option to configure it later.

* Fix error in NewFollower Event

An attempt to index a nil value error was reported on the devforum:
http://devforum.roblox.com/t/the-developer-console-core-script-bug-logs-they-burn/33356

Not sure where this error is coming from but adding this simple check
should resolve it.

* Check that sentNotificationMap exists before indexing it

Got a report of error stack trace referencing sentNotificationMap
showing up in a developers error log. Also changed it so the map is only
created for non-guest players similar to the PlayerToRelationshipMap.