CameraScripts optimizations changes
*Optimizations:*
- Make PopperCamera use
new Camera:GetLargestCutoffDistance() API, this is significantly faster
than the Lua version.
- Change PopperCamera to use new Camera:GetPartsObscuringTarget(), this
is faster than the Lua version.
- Change ordering of some if statements to do most common/less expensive
checks first. Then the short circuiting makes the check less expensive
on average.
- Change ClickToMove to use CFrame.lerp rather than Lua based
interpolation.
- Use FindFirstChildOfClass rather than iterating over children in a few
cases.

*Micro-optimizations:*
- Localize library functions.
- Don't construct new Vector3 and Vector2 values when it's not
necessary.
- Use t[#t + 1] instead of table.insert.
- Change in pairs loops to regular for loops where possible.

*Changes not included in this PR:*
- Using new Instance:GetPropertyChangedSignal() API instead of listening
to the Changed signal on the Camera object. I need to wait for the new
API to go live to add this.