CLIAVATAR-33: Fix shift lock allowing players to see through walls (#821)
This change goes along with a change to the implementation of
Camera:GetLargestCutoffDistance() which returns a maximum cutoff
distance if the camera focus is inside another part. If the focus moves
in another part this change pops the camera in fully and moves the focus
to the subject position rather than the shift lock offset position.